Proper accounting in setRecentInventoryKnown when we get a reject
If Node A sends us a tx and we reject it and then Node B also sends us
an INV for the same recently reject tx then we need to send a REJECT
notice to node B even though we will not attempt to download the txn.  This
is so that Node B will have a properly updated setRecentInventoryKnown in
the event that we request and XTHIN from them.