Parallel Block Validation to mitigate DDOS big block attack
- add 3 more script check queues
- use a wrapper function to call HandleBlockMessageThread
- Create semaphores to control and handle queue selection and throughput
- Update locking - we do not need to hold cs_main when checking inputs
- Flush the temporary view to the base view rather than re-running every txn with UpdateCoins
- unlock cs_main before waiting for script check threads to complete
- Python test for Parallel Validation

- Disable SENDHEADERS requests and always revert to INV
Seems that SENDHEADERS functionality really causes problems
even more so with parallel validation.  Turning it off here
completely, but we will still service nodes that want it.

- Only allow parallel block validation during IBD or receiving new blocks

- When mining new blocks or re-indexing or any other activity the blocks
are prevented from entering parallel validation: the cs_main locks are
not relinquished in those cases.

- Turn off Parallel Validation when doing IBD

- Ability to Quit scriptcheck threads

When  4 blocks are running in parallel and a
new, smaller block shows up, we need to be able to interrupt the
script threads that are currently validating for a one of the  blocks
so that we can free up a script check queue for the new smaller  block.

Documentation for Parallel Validation

Change some logprint messages from parallel to parallel_2

- Update the nBlockSequenceId after the block has advanced the tip

This is important for Parallel Validation.  By decreasing the sequence id
we are indicating that this block represents the current pindexMostWork. This
prevents the losing block from having the pindexMostWork point to it rather
than to the winning block.

- Continously check for new blocks to connect

With Parallel Validation new blocks can be accepted while we are connecting at
tip therefore we need to check after we connect each block whether a potentially
longer more work chain now exists. If we don't do this then we can at times end
up with a temporarily unconnected block until the next block gets mined and
another attempt is made to connect the most work chain.

- Terminate any PV threads during a re-org

PV can only operate with blocks that will advance the current chaintip (fork1). Therefore,
if we are needing to re-org to another chain (fork 2)  then we have to kill any currently
running PV threads assoicated with the current chain tip on fork1. This is the
solution to the problem of having two forks being mined while there is an attack
block processing on fork1. If fork2 continues to be mined and eventually pulls
ahead of fork1 in proof of work, then a re-org to fork2 will be initiated causing
the PV threads on fork1 to be terminated and fork2 blocks then connected and fork2 then
becoming the chain active tip.

- Use Chain Work instead of nHeight to self terminate a PV thread

If the Chain Work has changed either positve or negative to where it
was when we started the PV thread then we will exit the thread.  Previously
we were using Chain Height, which worked fine, but this is more understantable
from a coding perspective and also we added the feature to check for when
Chain Work has decreased from the starting point which would indicate that
a re-org was underway.

Move ZMQ notifications to ActivateBestChainStep

We must notify ZMQ after each tip is connected rather
than after ActivateBestChain otherwise we can miss a block
when there are several to connect together at once.

This particularly seems to help sync issues on regest where
we're mining many blocks all at the same time.