Move cs_blocksemaphore into globals.
Using it defined as a static within HandleBlockMessageThread is not
threadsafe until we get to c++11.

Move the aquistion of the script control until after the scoped lock

Change back to using a mutex for fQuit flag

It seems overkill to use atomic here for code that will
very rarely get executed.