Fix issue with Pruning
Files were not getting pruned which should have been because
the pruning function was using pindex->nHeight rather than the
current active chain height.  So when someone was syncing a new
chain in pruning mode the *.dat block files would often not
get unlinked as they should have.  If however the chain was
fully sync'd then the pruning would work just fine.  This
fix will allow new users to sync a chain from scratch in with
pruning turned on.