Request full block if thinblock re-construction fails
If the reconstruction is still missing any transactions for any reason
then request a full block.  (This should never happen because we
just checked the various pools for those transactions)

Save a little memory by clearing the mapPartialTxHash when no
longer needed.  In the event of a re-request we don't need