Add functions to determine the max system memory and use them
Use half the total system memory as the value for -dbcache
and use this value if the node operator has not already set
a nummber for -dbcache.  This way we can be pretty sure a node
will get a much better initial sync rather than in the past where
we would only rely on the default of 500MB for -dbcache.

Add getTotalSystemMemory() for UNIX and MACOS

Also return with a default value if physical memory size could
not be determined. We return with a default of 1GB which should
cover any raspberry pi configurations.

Make a more accurate calculation on Windows for memory available

On Windows systems we can know how much of the physical RAM is available
for use and so here we make a more fine grained calculation for the
amount of memory we'll use for the dbcache mechanism. For Windows systems
with large amounts of RAM available we'll use all of it up to the 16GB
maximum allowed for dbcaching, however, we'll always leave 1GB free
to the system for other uses.