Remove tracking of validated headers
This is not a useful metric anymore since we almost all the time,
except under unusual circumstances, will ask for a block without
having the header already. Furthermore, we were not tracking the data
correctly anyway.

More importantly, this would not allow us to finally remove cs_main
from MarkBlocksAsInFlight because of the need to track pindex. So
removing this is yet another step toward finally getting cs_main
out of the request manager.